Se modifico la invocacion al servicio, se agrego nuevo endpoint, pero es necesario hacer el levantamiento de los juegos y luego el recorrido de los items

+++ src/main/java/com/slackmag/domain/polla/Game.java
@@ -7,7 +7,7 @@ public class Game {
     public Integer gameCode;
     public Integer drawNumber;
     public Integer virtualDrawNumber;
-    public Integer deadline;
-    public Integer originalDeadline;
+    public Long deadline;
+    public Long originalDeadline;
 
 }
